Split the utilities in multiple modules
