+++ tsconfig.json
@@ -13,7 +13,7 @@
             "@store": ["@store"],
             "@types": ["@types"],
             "@constants": ["@constants"],
-            "@utils": ["@utils"],         
+            "@utils/*": ["@utils/*"],
             "@classes/*": ["@classes/*"]
         }
     },
